Attempt to fix pop in collision for high freqs
If boat is requesting wavelengths lower than currently rendering, when the lower LOD does start rendering it will pop in and can send the boat flying. This primarily happens when the camera changes altitude. This change kinda fixes it - it blends out the shape in LOD0. However the contents of LOD0 are already blended in the ocean shader, so this ends up with a power 2 blend that kills detail.