Attempt 2 to fix pop in collision for high freqs
This attempted to remove the aforementioned double blend of LOD0 - it still tranisitions out the geometry and blends out the normals, but does not blend out the shape. This seems to be working to some degree but causes weird pops in the mid ground. Moreover its getting so complicated I am struggling to keep all the rules in my head, so I'm parking this branch.